+++ tests/unit/test_builder.py
@@ -73,23 +73,37 @@ def test_initialize(monkeypatch: pytest.MonkeyPatch, tmp_path: Path):
     assert test_clouds_yaml_path.exists()
 
 
-def test__install_dependencies_fail(monkeypatch: pytest.MonkeyPatch):
+@pytest.mark.parametrize(
+    "apt_mock,pipx_install_mock",
+    [
+        pytest.param(
+            MagicMock(side_effect=apt.PackageNotFoundError("package not found")),
+            MagicMock(),
+            id="apt error",
+        ),
+        pytest.param(
+            MagicMock(),
+            MagicMock(side_effect=builder.PipXError("error installing deps")),
+            id="pipx install error",
+        ),
+    ],
+)
+def test__install_dependencies_fail(
+    apt_mock: MagicMock, pipx_install_mock, monkeypatch: pytest.MonkeyPatch
+):
     """
     arrange: monkeypatched subprocess.run function.
     act: when _install_dependencies is called.
     assert: DependencyInstallError is raised.
     """
-    monkeypatch.setattr(apt, "add_package", MagicMock())
-    monkeypatch.setattr(
-        pipx,
-        "install",
-        MagicMock(side_effect=PipXError("error installing deps")),
-    )
+    monkeypatch.setattr(apt, "add_package", apt_mock)
+    monkeypatch.setattr(pipx, "uninstall", MagicMock())
+    monkeypatch.setattr(pipx, "install", pipx_install_mock)
 
     with pytest.raises(builder.DependencyInstallError) as exc:
         builder._install_dependencies()
 
-    assert "error installing deps" in str(exc.getrepr())
+    assert exc.value.__cause__ == (apt_mock.side_effect or pipx_install_mock.side_effect)
 
 
 def test__install_dependencies(monkeypatch: pytest.MonkeyPatch):
@@ -99,12 +113,39 @@ def test__install_dependencies(monkeypatch: pytest.MonkeyPatch):
     assert: mocked functions are called.
     """
     monkeypatch.setattr(apt, "add_package", (apt_mock := MagicMock()))
-    monkeypatch.setattr(pipx, "install", (install_mock := MagicMock()))
+    monkeypatch.setattr(pipx, "uninstall", (pipx_uninstall_mock := MagicMock()))
+    monkeypatch.setattr(pipx, "install", (pipx_install_mock := MagicMock()))
 
     builder._install_dependencies()
 
     apt_mock.assert_called_once()
-    install_mock.assert_called_once()
+    pipx_uninstall_mock.assert_called_once()
+    pipx_install_mock.assert_called_once()
+
+
+def test__install_dependencies_uninstalling_app_fail(
+    monkeypatch: pytest.MonkeyPatch, caplog: pytest.LogCaptureFixture
+):
+    """
+    arrange: given monkeypatched pipx.uninstall function that raises an error.
+    act: when _install_dependencies is called.
+    assert: the error is ignored and an information is logged.
+    """
+    monkeypatch.setattr(apt, "add_package", (apt_mock := MagicMock()))
+    monkeypatch.setattr(
+        pipx,
+        "uninstall",
+        (pipx_uninstall_mock := MagicMock(side_effect=PipXError("error uninstalling deps"))),
+    )
+    monkeypatch.setattr(pipx, "install", (pipx_install_mock := MagicMock()))
+
+    builder._install_dependencies()
+
+    apt_mock.assert_called_once()
+    pipx_uninstall_mock.assert_called_once()
+    pipx_install_mock.assert_called_once()
+
+    assert "error uninstalling deps" in " ".join(caplog.messages)
 
 
 @pytest.mark.parametrize(
@@ -348,8 +389,8 @@ def test_configure_cron_no_reconfigure(monkeypatch: pytest.MonkeyPatch):
     "expected_file_contents",
     [
         pytest.param(
-            """0 */1 * * * ubuntu /usr/bin/run-one /usr/bin/bash -c \
-/usr/bin/juju-exec "test-unit-name" "JUJU_DISPATCH_PATH=run HOME=/home/ubuntu ./dispatch"
+            """0 */1 * * * root /usr/bin/run-one /usr/bin/bash -c \
+'/usr/bin/juju-exec "test-unit-name" "JUJU_DISPATCH_PATH=run HOME=/home/ubuntu ./dispatch"'
 """,
             id="runner version set",
         ),
@@ -929,49 +970,3 @@ def test__get_latest_image(monkeypatch: pytest.MonkeyPatch):
         cloud_id="test-cloud",
         image_id="test-image",
     )
-
-
-def test_upgrade_app(monkeypatch: pytest.MonkeyPatch):
-    """
-    arrange: given monkeypatched pipx commands.
-    act: when upgrade_app is called.
-    assert: expected pipx commands are run.
-    """
-    monkeypatch.setattr(pipx, "uninstall", MagicMock())
-    monkeypatch.setattr(pipx, "install", MagicMock())
-
-    builder.upgrade_app()
-
-    # pylint does not recognize that these are MagicMock objects
-    pipx.uninstall.assert_called_once()  # pylint: disable=no-member
-    pipx.install.assert_called_once()  # pylint: disable=no-member
-
-
-@pytest.mark.parametrize(
-    "fct",
-    [
-        pytest.param(
-            "uninstall",
-            id="uninstall error",
-        ),
-        pytest.param(
-            "install",
-            id="install error",
-        ),
-    ],
-)
-def test_upgrade_app_error(
-    monkeypatch: pytest.MonkeyPatch,
-    fct: str,
-):
-    """
-    arrange: given monkeypatched subprocess.run that raises an error.
-    act: when upgrade_app is called.
-    assert: UpgradeApplicationError is raised.
-    """
-    monkeypatch.setattr(pipx, "uninstall", MagicMock())
-    monkeypatch.setattr(pipx, "install", MagicMock())
-    monkeypatch.setattr(pipx, fct, MagicMock(side_effect=PipXError))
-
-    with pytest.raises(builder.UpgradeApplicationError):
-        builder.upgrade_app()